[REF] Create new stats module (#273)
* Move statistical functions into new stats module.

* Update API.

* Requisite bug fix.

* Requisite *second* bug fix.
